% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{DEGAnalysis_DESeq2}
\alias{DEGAnalysis_DESeq2}
\title{DEG analysis using samples with replicates.}
\usage{
DEGAnalysis_DESeq2(
  countMatrix,
  group,
  useFDR = F,
  cut = 0.05,
  FCcut = 2,
  control = "Manually_select",
  treat = "Manually_select",
  outdir = NULL
)
}
\arguments{
\item{countMatrix}{input count matrix, rows are genes and columns are samples}

\item{group}{group dataframe, first column: sample ID, second column: group name}

\item{useFDR}{whether to use FDR to define significant DEG}

\item{cut}{significant P value threshold, is useFDR is set , FDR is used}

\item{FCcut}{Fold Change threshold}

\item{control}{control group name, used as denominator in Fold Change calculation}

\item{treat}{treat group name, used as numerator  in Fold Change calculation}

\item{outdir}{output file directory}
}
\description{
do DEG analysis using samples with replicates by DESeq2.
}
