% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{DEGAnalysis_EBSeq}
\alias{DEGAnalysis_EBSeq}
\title{DEG analysis using samples without replicates.}
\usage{
DEGAnalysis_EBSeq(
  countMatrix,
  cut = 0.05,
  FCcut = 2,
  outdir = NULL,
  control = "Manually_select",
  treat = "Manually_select"
)
}
\description{
do DEG analysis using samples without replicates by EBSeq.
}
