% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOenrich_common}
\alias{GOenrich_common}
\title{GO enrichment using  annotation.}
\usage{
GOenrich_common(
  deglist,
  geneID2GO,
  fdr = FALSE,
  outdir = NULL,
  outprefix = NULL
)
}
\description{
do GO enrichment by a list of DEGs using  annotation.
}
