% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOEnrich_eggnog}
\alias{GOEnrich_eggnog}
\title{GO enrichment using eggnog Mapper annotation.}
\usage{
GOEnrich_eggnog(
  deglist,
  alllist = NULL,
  eggnogFile,
  useFDR = FALSE,
  cut = 0.05,
  outdir = NULL,
  outprefix = NULL
)
}
\arguments{
\item{deglist}{DEG vector}

\item{eggnogFile}{custom annotation file using eggNOG database}

\item{useFDR}{whether to use FDR}

\item{cut}{significant pvalue threshold, if useFDR is set, use FDR instead}

\item{outdir}{output directory}

\item{outprefix}{output file prefix}
}
\description{
do GO enrichment by a list of DEGs using eggnog Mapper annotation.
}
