% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{KEGGbubble}
\alias{KEGGbubble}
\title{KEGG enrichment bubble plot.}
\usage{
KEGGbubble(
  dataset,
  MainTitle = "",
  TrimName = F,
  NameWidth = 50,
  top = 10,
  ColorScheme = "rw",
  ColorReverse = F,
  xangle = 0
)
}
\arguments{
\item{dataset}{a dataframe of KEGG Enrichment output, with these columns:'ID', 'KEGGPathway', 'Count', 'all', 'P'}

\item{MainTitle}{main title}

\item{top}{top n kegg pathways used to plot}

\item{ColorScheme}{"ryb", 'rwb', 'ryg', 'rwg', 'rw', 'bw', 'gw'}

\item{ColorReverse}{whether to reverse color order}

\item{xangle}{angle of x axis text}
}
\description{
do KEGG pathway enrichment bubble plot using KEGGenrich result.
}
