% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{KEGGenrich_customTable}
\alias{KEGGenrich_customTable}
\title{KEGG pathway enrichment tools using custom annotation file.}
\usage{
KEGGenrich_customTable(
  deglist,
  alllist = NULL,
  tableFile,
  outdir = NULL,
  outprefix,
  cutP = 0.05,
  useFDR = FALSE
)
}
\arguments{
\item{deglist}{DEG vector}

\item{tableFile}{custom kegg to map file}

\item{outdir}{output directory}

\item{outprefix}{output file prefix}

\item{useFDR}{whether to add FDR column}
}
\description{
do KEGG pathway enrichment by a list of degs using custom annotation file.
}
