% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{MAPlot}
\alias{MAPlot}
\title{MA plot for DEG analysis results.}
\usage{
MAPlot(
  dat,
  useFDR = T,
  MainTitle = "",
  cut = 0.05,
  FCcut = 2,
  ylim = 10,
  showlabel = "no",
  showlabel.num = 10,
  point.size = 2
)
}
\description{
do MA plot.
}
