% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{matrixGroup}
\alias{matrixGroup}
\title{merge column values according group information.}
\usage{
matrixGroup(matrix, groupInfo, method = "mean")
}
\description{
merge column values according group information.
}
