% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{multiTraitPlot}
\alias{multiTraitPlot}
\title{trait Plot.}
\usage{
multiTraitPlot(data)
}
\arguments{
\item{data}{input trait dataframe, with traits as columns}
}
\description{
Trait plot with histograms/correlation/density.
}
