% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{multiTraitStat}
\alias{multiTraitStat}
\title{Frequently used stats for trait analysis.}
\usage{
multiTraitStat(data)
}
\arguments{
\item{data}{input trait dataframe, with traits as columns}
}
\description{
Frequently used stats for trait analysis.
}
