% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{PANNZERres2GOdb}
\alias{PANNZERres2GOdb}
\title{convert PANNZER GO annotation to godb addon.}
\usage{
PANNZERres2GOdb(PANNZERfile)
}
\description{
convert PANNZER GO annotation(GO.out file) to geneID2GO format.
}
