% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Volcano plot for DEG analysis results.}
\usage{
VolcanoPlot(
  dat,
  useFDR = T,
  MainTitle = "",
  cut = 0.05,
  FCcut = 2,
  xlim = 10,
  showlabel = "no",
  showlabel.num = 10,
  point.size = 2
)
}
\arguments{
\item{dat}{deg analysis result dataframe with ID,FCcut,log2FoldChange,Pvalue or FDR}

\item{useFDR}{whether to use FDR to define significance}

\item{MainTitle}{main title of plot}

\item{cut}{P value significant cut , if useFDR is set, FDR is used}

\item{FCcut}{FoldChange cut significant cut}

\item{xlim}{x axis limits}

\item{showlabel}{whether to show top significant gene names}

\item{showlabel.num}{top number}

\item{point.size}{point size}
}
\description{
do volcano plot.
}
