% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{DEGAnalysis_DESeq2}
\alias{DEGAnalysis_DESeq2}
\title{DEG analysis using samples with replicates.}
\usage{
DEGAnalysis_DESeq2(
  countMatrix,
  group,
  useFDR = F,
  cut = 0.05,
  FCcut = 2,
  control = NULL,
  treat = NULL,
  outdir = NULL
)
}
\description{
do DEG analysis using samples with replicates by DESeq2.
}
