% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{GObubble}
\alias{GObubble}
\title{GO enrichment bubble plot.}
\usage{
GObubble(
  dataset,
  MainTitle = "",
  onlySig = T,
  pCut = 0.05,
  top = 10,
  ColorScheme = "rw",
  ColorReverse = F
)
}
\description{
do GO  enrichment bubble plot using GOEnrich result.
}
