% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOEnrich_customMapping}
\alias{GOEnrich_customMapping}
\title{GO enrichment using custom mapping file.}
\usage{
GOEnrich_customMapping(
  deglist,
  mappingfile,
  fdr = FALSE,
  outdir = NULL,
  outprefix = NULL
)
}
\description{
do GO enrichment by a list of DEGs using custom mapping  file.
}
