% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOEnrich_customMapping}
\alias{GOEnrich_customMapping}
\title{GO enrichment using custom mapping file.}
\usage{
GOEnrich_customMapping(
  deglist,
  mappingfile,
  useFDR = FALSE,
  cut = 0.05,
  outdir = NULL,
  outprefix = NULL
)
}
\arguments{
\item{deglist}{DEG vector}

\item{mappingfile}{path of the custom GO annotation file, which annotates each gene in a row, separate GOs with comma(,)}

\item{useFDR}{whether to use FDR}

\item{cut}{significant pvalue threshold, if useFDR is set, use FDR instead}

\item{outdir}{output directory}

\item{outprefix}{output file prefix}
}
\description{
do GO enrichment by a list of DEGs using custom mapping  file.
}
