% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOEnrich_customTable}
\alias{GOEnrich_customTable}
\title{GO enrichment using custom table  annotation.}
\usage{
GOEnrich_customTable(
  deglist,
  tableFile,
  fdr = FALSE,
  outdir = NULL,
  outprefix = NULL
)
}
\description{
do GO enrichment by a list of DEGs using custom table annotation.
}
