% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOEnrich_customTable}
\alias{GOEnrich_customTable}
\title{GO enrichment using custom table  annotation.}
\usage{
GOEnrich_customTable(
  deglist,
  tableFile,
  useFDR = FALSE,
  cut = 0.05,
  outdir = NULL,
  outprefix = NULL
)
}
\arguments{
\item{deglist}{DEG vector}

\item{tableFile}{path of the custom GO annotation file, which annotates two-column pairwise gene-GO per row}

\item{useFDR}{whether to use FDR}

\item{cut}{significant pvalue threshold, if useFDR is set, use FDR instead}

\item{outdir}{output directory}

\item{outprefix}{output file prefix}
}
\description{
do GO enrichment by a list of DEGs using custom table annotation.
}
