% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOEnrich_eggnog}
\alias{GOEnrich_eggnog}
\title{GO enrichment using eggnog Mapper annotation.}
\usage{
GOEnrich_eggnog(
  deglist,
  eggnogFile,
  fdr = FALSE,
  outdir = NULL,
  outprefix = NULL
)
}
\description{
do GO enrichment by a list of DEGs using eggnog Mapper annotation.
}
