% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{GOEnrich_pannzer2}
\alias{GOEnrich_pannzer2}
\title{GO enrichment using pannzer2 result.}
\usage{
GOEnrich_pannzer2(
  deglist,
  pannzerfile,
  fdr = FALSE,
  outdir = NULL,
  outprefix = NULL
)
}
\description{
do GO enrichment by a list of DEGs using pannzer2 result.
}
