% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{KEGGbubble}
\alias{KEGGbubble}
\title{KEGG enrichment bubble plot.}
\usage{
KEGGbubble(
  dataset,
  MainTitle = "",
  top = 10,
  ColorScheme = "rw",
  ColorReverse = F,
  xangle = 45
)
}
\description{
do KEGG pathway enrichment bubble plot using KEGGenrich result.
}
