% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{KEGGenrich}
\alias{KEGGenrich}
\title{KEGG pathway enrichment tools.}
\usage{
KEGGenrich(deglist, taxon, outdir = NULL, outprefix, useFDR = FALSE)
}
\description{
do KEGG pathway enrichment by a list of degs.
}
