% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{KEGGenrich_blastkoala}
\alias{KEGGenrich_blastkoala}
\title{KEGG pathway enrichment tools for blastkoala ko annotation.}
\usage{
KEGGenrich_blastkoala(
  deglist,
  blastkoalafile,
  taxonid = "ko",
  outdir = NULL,
  outprefix,
  useFDR = FALSE
)
}
\description{
do KEGG pathway enrichment by a list of degs. Using blastkoala method to convert from GeneID to KEGG ID.
}
