% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.r
\name{KEGGenrich_common}
\alias{KEGGenrich_common}
\title{KEGG pathway enrichment tools commmon.}
\usage{
KEGGenrich_common(deglist, pathinfo, outdir = NULL, outprefix, fdr = FALSE)
}
\description{
do KEGG pathway enrichment by a list of degs common.
}
