% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Volcano plot for DEG analysis results.}
\usage{
VolcanoPlot(
  dat,
  useFDR = T,
  MainTitle = "",
  cut = 0.05,
  FCcut = 2,
  xlim = 10,
  showlabel = "no",
  showlabel.num = 10,
  point.size = 2
)
}
\description{
do volcano plot.
}
